Added in police districts
@semerj at this point things are getting hairy, I'm not sure what we
should be completely doing but we should really focus on telling a
story. I'd like to try and pick one and go for it. I think that we
either tell a alcohol related story or we try and do some gentrification
of the mission district.

The page format is good for developing on.